# backend/project/endpoints/index/OpenAPI_Object.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pigeonhole API",
    "summary": "A project submission and grading API for University Ghent students and professors.",
    "description": "The API built for the Pigeonhole application. It serves as an interface for student of University Ghent. They can submit solutions to projects created by their professors. Professors and their assistents can then review these submitions, grade them and define custom tests that automatically run on every submition. The API is built using the OpenAPI 3.1.0 specification.",
    "version": "1.0.0",
    "contact": {
      "name": "Project discussion forum",
      "url": "https://github.com/SELab-2/UGent-opgave/discussions",
      "email": "Bart.Coppens@UGent.be"
    },
    "x-authors": [
      {
        "name": "Aron Buzogany",
        "github": "https://github.com/AronBuzogany"
      },
      {
        "name": "Gerwoud Van den Eynden",
        "github": "https://github.com/Gerwoud"
      },
      {
        "name": "Jarne Clauw",
        "github": "https://github.com/JarneClauw"
      },
      {
        "name": "Siebe Vlietinck",
        "github": "https://github.com/Vucis"
      },
      {
        "name": "Warre Provoost",
        "github": "https://github.com/warreprovoost"
      },
      {
        "name": "Cedric Mekeirle",
        "github": "https://github.com/JibrilExe"
      },
      {
        "name": "Matisse Sulzer",
        "github": "https://github.com/Matisse-Sulzer"
      }
    ]
  },
  "paths": {
...
    "/courses": {
      "get": {
        "description": "Get a list of all courses.",
        "responses": {
          "200": {
            "description": "Successfully retrieved all courses with given parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "url": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
            },
          "parameters": [
            {
              "name": "name",
              "in": "query",
              "description": "Name of the course",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "ufora_id",
              "in": "query",
              "description": "Ufora ID of the course",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "teacher",
              "in": "query",
              "description": "Teacher of the course",
              "schema": {
                "type": "string"
              }
            }
          ]
        },
      "post": {
        "description": "Create a new course.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the course"
                  },
                  "teacher": {
                    "type": "string",
                    "description": "Teacher of the course"
                  }
                },
                "required": ["name", "teacher"]
              }
            }
          }
        },
        "parameters":[
          {
            "name":"uid",
            "in":"query",
            "description":"uid of the user sending the request",
            "schema":{
              "type":"string"
            }
          }
        ],
        "responses":{
          "400": {
            "description": "There was no uid in the request query.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "Course with name: {name} and course_id: {course_id} was successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "course_id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "teacher": {
                          "type": "string"
                        },
                        "ufora_id": {
                          "type": "string"
                        }
                      }
                    },
                    "url": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "The user trying to create a course was unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "The user trying to create a course was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }},
...
    "/users": {
      "get": {
        "summary": "Get all users",
        "responses": {
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "uid": {
                        "type": "string"
                      },
                      "is_teacher": {
                        "type": "boolean"
                      },
                      "is_admin": {
                        "type": "boolean"
                      }
                    },
                    "required": ["uid", "is_teacher", "is_admin"]
                  }
                }
              }
            }
          }}},
          "post": {
            "summary": "Create a new user",
            "requestBody": {
              "required": true,
              "content":{
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "uid": {
                        "type": "string"
                      },
                      "is_teacher": {
                        "type": "boolean"
                      },
                      "is_admin": {
                        "type": "boolean"
                      }
                    },
                    "required": ["uid", "is_teacher", "is_admin"]
                  }
                }
              }
            },
            "responses": {
              "201": {
                "description": "User created successfully"
              },
              "400": {
                "description": "Invalid request data"
              },
              "415": {
                "description": "Unsupported Media Type. Expected JSON."
              },
              "500": {
                "description": "An error occurred while creating the user"
              }
            }

        },
        "/users/{user_id}": {
          "get": {
            "summary": "Get a user by ID",
            "parameters": [
              {
                "name": "user_id",
                "in": "path",
                "required": true,
                "schema": {
                  "type": "string"
                }
              }
            ],
            "responses": {
              "200": {
                "description": "A user",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "uid": {
                          "type": "string"
                        },
                        "is_teacher": {
                          "type": "boolean"
                        },
                        "is_admin": {
                          "type": "boolean"
                        }
                      },
                      "required": ["uid", "is_teacher", "is_admin"]
                    }
                  }
                }
              },
              "404": {
                "description": "User not found"
              }
            }
          },
          "patch": {
            "summary": "Update a user's information",
            "parameters": [
              {
                "name": "user_id",
                "in": "path",
                "required": true,
                "schema": {
                  "type": "string"
                }
              }
            ],
            "requestBody": {
              "required": true,
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "is_teacher": {
                        "type": "boolean"
                      },
                      "is_admin": {
                        "type": "boolean"
                      }
                    },
                    "required": ["is_teacher", "is_admin"]
                  }
                }
              }
            },
            "responses": {
              "200": {
                "description": "User updated successfully"
              },
              "404": {
                "description": "User not found"
              },
              "415": {
                "description": "Unsupported Media Type. Expected JSON."
              },
              "500": {
                "description": "An error occurred while patching the user"
              }
            }
          },
          "delete": {
            "summary": "Delete a user",
            "parameters": [
              {
                "name": "user_id",
                "in": "path",
                "required": true,
                "schema": {
                  "type": "string"
                }
              }
            ],
            "responses": {
              "200": {
                "description": "User deleted successfully"
              },
              "404": {
                "description": "User not found"
              },
              "500": {
                "description": "An error occurred while deleting the user"
              }
            }
          }
        }
      }
  }
}